# api-docs/api-doc.yml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:61097
  description: Generated server url
...
paths:
  /{version}/merchants/{merchantId}/products/{productId}:
    get:
      tags:
      - Merchant Products
      summary: Get a single product
      operationId: get
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: branchId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
    put:
      tags:
      - Merchant Products
      summary: "Update a product's name, price, and image"
      operationId: update
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                branchId:
                  type: string
                unitPrice:
                  type: string
                description:
                  type: string
                categoryId:
                  type: string
                productCode:
                  type: string
                image:
                  type: string
                  format: binary
              required:
              - branchId
              - name
              - unitPrice
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
    delete:
      tags:
      - Merchant Products
      summary: Soft-delete a product (sets active = false)
      operationId: deactivate
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: branchId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
...
  /{version}/merchants/{merchantId}/location:
    get:
      tags:
      - Merchant Location
      summary: Get a merchant's current location
      description: Returns 404 if the merchant has no location set.
      operationId: getLocation
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MerchantLocationResponse"
    put:
      tags:
      - Merchant Location
      summary: Set or update a merchant's location and geofence radius
      description: Idempotent — replaces any existing location. geofenceRadiusMeters
        must be > 0. googlePlaceId is optional and used to correlate with Google Maps.
      operationId: setLocation
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetMerchantLocationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MerchantLocationResponse"
    delete:
      tags:
      - Merchant Location
      summary: Remove a merchant's location
      description: "Clears latitude, longitude, geofence radius, and googlePlaceId.\
        \ Returns 204 on success."
      operationId: deleteLocation
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /{version}/merchants/{merchantId}/branches/{branchId}/image:
    put:
      tags:
      - Merchant Branches
      summary: Set branch image
      operationId: setImage
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: branchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetBranchImageRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BranchResponse"
...
  /{version}/merchants/{id}:
    get:
      tags:
      - Merchant Onboarding
      summary: Get merchant application details
      operationId: get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MerchantDetailsResponse"
    put:
      tags:
      - Merchant Onboarding
      summary: Update editable merchant details and optional status
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMerchantRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MerchantDetailsResponse"
...
  /{version}/merchants/nearby:
    get:
      tags:
      - Merchant Location
      summary: Find merchants near a position
      description: "Returns merchants that have at least one active branch location\
        \ within radiusMeters of (lat, lng). Intended for mobile geofence-enter events.\
        \ Default search radius is 500 m. Results are ordered by ascending nearest\
        \ branch distance. Merchant-level location is ignored."
      operationId: findNearby
      parameters:
      - name: lat
        in: query
        required: true
        schema:
          type: number
          format: double
      - name: lng
        in: query
        required: true
        schema:
          type: number
          format: double
      - name: radiusMeters
        in: query
        required: false
        schema:
          type: number
          format: double
          default: 500.0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NearbyMerchantsResponse"
...
components:
  schemas:
    ProductResponse:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        branchId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        unitPrice:
          type: number
        imageUrl:
          type: string
        active:
          type: boolean
        status:
          type: string
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        productCode:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - active
      - branchId
      - createdAt
      - name
      - productId
      - status
      - unitPrice
      - updatedAt
...
    SetMerchantLocationRequest:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        geofenceRadiusMeters:
          type: integer
          format: int32
        googlePlaceId:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
      required:
      - geofenceRadiusMeters
      - latitude
      - longitude
    MerchantLocationResponse:
      type: object
      properties:
        merchantId:
          type: string
          format: uuid
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        geofenceRadiusMeters:
          type: integer
          format: int32
        googlePlaceId:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        placeFormattedAddress:
          type: string
        placeRating:
          type: number
          format: double
        placeReviewCount:
          type: integer
          format: int32
        placeEnrichmentStatus:
          type: string
        placeEnrichedAt:
          type: string
          format: date-time
      required:
      - geofenceRadiusMeters
      - latitude
      - longitude
      - merchantId
    SetBranchImageRequest:
      type: object
      properties:
        storageKey:
          type: string
      required:
      - storageKey
    BranchResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchantId:
          type: string
          format: uuid
        name:
          type: string
        active:
          type: boolean
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        contactFullName:
          type: string
        contactEmail:
          type: string
        contactPhoneNumber:
          type: string
        activePaymentChannels:
          type: array
          items:
            type: string
            enum:
            - PAYPAL
            - STRIPE
            - TRUELAYER
          uniqueItems: true
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        geofenceRadiusMeters:
          type: integer
          format: int32
        googlePlaceId:
          type: string
        placeFormattedAddress:
          type: string
        placeRating:
          type: number
          format: double
        placeReviewCount:
          type: integer
          format: int32
        placeEnrichmentStatus:
          type: string
        placeEnrichedAt:
          type: string
          format: date-time
        imageUrl:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - active
      - activePaymentChannels
      - createdAt
      - id
      - merchantId
      - name
      - status
      - updatedAt
...
    UpdateMerchantRequest:
      type: object
      properties:
        legalBusinessName:
          type: string
        primaryBusinessAddress:
          type: string
        contactFullName:
          type: string
        contactEmail:
          type: string
        contactPhoneNumber:
          type: string
        activePaymentChannels:
          type: array
          items:
            type: string
            enum:
            - PAYPAL
            - STRIPE
            - TRUELAYER
          uniqueItems: true
        status:
          type: string
          enum:
          - DRAFT
          - SUBMITTED
          - VERIFICATION
          - SCREENING
          - RISK_REVIEW
          - DECISION_PENDING
          - SETUP
          - ACTIVE
          - MONITORING
          - ACTION_REQUIRED
          - REJECTED
      required:
      - activePaymentChannels
      - contactEmail
      - contactFullName
      - contactPhoneNumber
      - legalBusinessName
      - primaryBusinessAddress
    MerchantDetailsResponse:
      type: object
      properties:
        applicationId:
          type: string
          format: uuid
        applicationReference:
          type: string
        registrationNumber:
          type: string
        businessType:
          type: string
        operatingCountry:
          type: string
        legalBusinessName:
          type: string
        primaryBusinessAddress:
          type: string
        contactFullName:
          type: string
        contactEmail:
          type: string
        contactPhoneNumber:
          type: string
        activePaymentChannels:
          type: array
          items:
            type: string
            enum:
            - PAYPAL
            - STRIPE
            - TRUELAYER
          uniqueItems: true
        status:
          type: string
          enum:
          - DRAFT
          - SUBMITTED
          - VERIFICATION
          - SCREENING
          - RISK_REVIEW
          - DECISION_PENDING
          - SETUP
          - ACTIVE
          - MONITORING
          - ACTION_REQUIRED
          - REJECTED
        submittedAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
        logoStorageKey:
          type: string
        logoUrl:
          type: string
      required:
      - activePaymentChannels
      - applicationId
      - applicationReference
      - businessType
      - contactEmail
      - contactFullName
      - contactPhoneNumber
      - lastUpdatedAt
      - legalBusinessName
      - operatingCountry
      - primaryBusinessAddress
      - registrationNumber
      - status
...
    NearbyBranchResponse:
      type: object
      properties:
        branchId:
          type: string
          format: uuid
        name:
          type: string
        distanceMeters:
          type: number
          format: double
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        addressLine1:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        contactFullName:
          type: string
        contactEmail:
          type: string
        contactPhoneNumber:
          type: string
        activePaymentChannels:
          type: array
          items:
            type: string
            enum:
            - PAYPAL
            - STRIPE
            - TRUELAYER
          uniqueItems: true
      required:
      - activePaymentChannels
      - branchId
      - name
    NearbyMerchantResponse:
      type: object
      properties:
        merchantId:
          type: string
          format: uuid
        legalBusinessName:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        geofenceRadiusMeters:
          type: integer
          format: int32
        googlePlaceId:
          type: string
        distanceMeters:
          type: number
          format: double
        activeBranches:
          type: array
          items:
            $ref: "#/components/schemas/NearbyBranchResponse"
      required:
      - activeBranches
      - distanceMeters
      - geofenceRadiusMeters
      - latitude
      - legalBusinessName
      - longitude
      - merchantId
    NearbyMerchantsResponse:
      type: object
      properties:
        merchants:
          type: array
          items:
            $ref: "#/components/schemas/NearbyMerchantResponse"
      required:
      - merchants
...